Implement workload update

+++ src/Cli/dotnet/commands/dotnet-workload/install/NetSdkManagedInstaller.cs
@@ -15,6 +15,7 @@ using NuGet.Common;
 using NuGet.Versioning;
 using static Microsoft.NET.Sdk.WorkloadManifestReader.WorkloadResolver;
 using Microsoft.DotNet.Workloads.Workload.Install.InstallRecord;
+using System.Text.Json;
 
 namespace Microsoft.DotNet.Workloads.Workload.Install
 {
@@ -269,6 +270,13 @@ namespace Microsoft.DotNet.Workloads.Workload.Install
                         unneededBandRecords = unneededBandRecords.Append(currentBandRecordPath);
                     }
 
+                    if (!unneededBandRecords.Any())
+                    {
+                        continue;
+                    }
+
+                    // Save the pack info in case we need to delete the pack
+                    var jsonPackInfo = File.ReadAllText(unneededBandRecords.First());
                     foreach (var unneededRecord in unneededBandRecords)
                     {
                         File.Delete(unneededRecord);
@@ -278,6 +286,11 @@ namespace Microsoft.DotNet.Workloads.Workload.Install
                     {
                         Directory.Delete(packVersionDir);
                         var deletablePack = GetPackInfo(packVersionDir);
+                        if (deletablePack == null)
+                        {
+                            // Pack no longer exists in manifests, get pack info from installation record
+                            deletablePack = JsonSerializer.Deserialize(jsonPackInfo, typeof(PackInfo)) as PackInfo;
+                        }
                         DeletePack(deletablePack);
                     }
                 }
@@ -356,7 +369,7 @@ namespace Microsoft.DotNet.Workloads.Workload.Install
             {
                 Directory.CreateDirectory(Path.GetDirectoryName(path));
             }
-            File.Create(path);
+            File.WriteAllText(path, JsonSerializer.Serialize(packInfo));
         }
 
         private void DeletePackInstallationRecord(PackInfo packInfo, SdkFeatureBand featureBand)
